[cleanup] Move TableNav tests to table.rs.
It was in app.rs where TableNav originally located.
